Rewrote the way of working with a group of tasks
Remove enums (badly scaled and redundant) and replace enums to
MTTaskGroup objects.
Now you can create as many groups as you need and give them names as you
wish.

+++ Scheduler/Source/MTMicroWebSrv.cpp
@@ -225,7 +225,7 @@ void MicroWebServer::Append(const char * txt)
 	}
 
 	memcpy(answerData + answerSize, txt, stringLen);
-	answerSize += stringLen;
+	answerSize += (uint32)stringLen;
 	answerData[answerSize] = '\0';
 }
